+++ clang/test/Driver/constructors.c
@@ -34,7 +34,7 @@
 // RUN:     -target i386-unknown-linux \
 // RUN:     --sysroot=%S/Inputs/basic_linux_tree \
 // RUN:     --gcc-toolchain="" \
-// RUN:   | FileCheck --check-prefix=CHECK-NO-INIT-ARRAY %s
+// RUN:   | FileCheck --check-prefix=CHECK-INIT-ARRAY %s
 //
 // RUN: %clang -no-canonical-prefixes %s -### -fsyntax-only 2>&1       \
 // RUN:     -fuse-init-array \